UI update search input
