+++ src/Components/AdoptionForms/AdoptionForms.scss
@@ -2,11 +2,10 @@
 @use "../../styles/partials/variables" as *;
 
 .form {
-    display: flex;
-    flex-direction: column;
+  display: flex;
+  flex-direction: column;
   gap: 1rem;
 }
-
 
 .form__information--container {
   display: flex;
@@ -14,39 +13,5 @@
   gap: 1rem;
   @include tablet {
     flex-direction: row;
-  }
-}
-
-.searchInput__container {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  border: 1px solid #ccc;
-  border-radius: 1rem;
-  padding: 0 1rem;
-  background-color: #eef2f5;
-}
-
-.searchInput__container input[type="text"] {
-  padding: 0.7rem;
-  font-size: 16px;
-  width: 300px;
-  margin-right: 10px;
-  background-color: #eef2f5;
-  border: none;
-  &:focus {
-    outline: none;
-  }
-  &::placeholder {
-    color: #666;
-  }
-}
-
-.error {
-  color: red;
-  font-size: small;
-  padding-top: 0.2rem;
-  &__border {
-    border: 1px solid red;
   }
 }
