+++ src/Components/Button/Button.scss
@@ -5,10 +5,13 @@
   width: 100%;
   height: 3rem;
   border: 1px solid #ccc;
-  border-radius: 1rem;
+  border-radius: 0.5rem;
   padding: 0.5rem;
   background-color: #070808;
   color: #fff;
+  display: flex;
+  justify-content: center;
+  align-items: center;
   @include tablet {
     padding: 0.7rem 2rem;
