+++ src/Components/Footer/Footer.scss
@@ -7,7 +7,7 @@
   flex-shrink: 0;
   padding: 1rem;
   @include tablet {
-    padding: 6rem 2rem;
+    padding: 2rem 2rem;
   }
   &__logo-container--text {
     font-size: 1.5rem;
