+++ src/Components/SearchInput/SearchInput.scss
@@ -1,25 +1,45 @@
-.searchInput__container {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  border: 1px solid #ccc;
-  border-radius: 1rem;
-  padding: 0 1rem;
-  background-color: #eef2f5;
-}
+@use "../../styles/partials/mixin" as *;
+@use "../../styles/partials/variables" as *;
 
-.searchInput__container input[type="text"] {
-  padding: 0.7rem;
-  font-size: 16px;
-  // widt  h: 100%;
-  margin-right: 10px;
-  background-color: #eef2f5;
-  border: none;
-  &:focus {
-    outline: none;
+.searchInput__container {
+  gap: 0.5rem;
+  // border: 1px solid #ccc;
+  border-radius: 1rem;
+  background-color: #141414;
+  width: 100%;
+  height: 10rem;
+  padding: 0.5rem 1rem 0 0.5rem;
+
+  @include tablet {
+    width: 30rem;
+    height: 10rem;
   }
-  &::placeholder {
-    color: #666;
+  &__text--container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    @include tablet {
+      gap: 2rem;
+    }
+    & textarea {
+      flex: 1;
+      // padding: 1rem 1rem 0 1rem;
+      height: 100%;
+      padding: 12px 20px;
+      box-sizing: border-box;
+      border-radius: 1rem;
+      font-size: 16px;
+      resize: none;
+      background-color: #141414;
+      border: none;
+      &:focus {
+        outline: none;
+      }
+      &::placeholder {
+        color: white;
+        font-weight: 600;
+      }
+    }
   }
 }
